requirements.txt
streamlit
PyMuPDF
matplotlib
scikit-learn
pytesseract
pdf2image
Pillow
transformers
sentence-transformers
keybert
spacy
pdfplumber
python-docx
numpy
captcha
python-dotenv
PyPDF2
